Establish means for including common lists
I think my content will call on some common lists, possibly in
different configurations under different circumstances. This is not yet
implemented, but it does work. The point is to have a document with
tagged includable sections for parent docs to pull in; those includes
in turn generated from semi-structured source, such as item-lists.yml
here